:truck: rename prisma file to models
