+++ src/utils/index.ts
@@ -2,8 +2,7 @@ export * from './cn';
 export * from './constants';
 export * from './date';
 export * from './functions';
-export * from './gcp';
 export * from './limits';
-export * from './prisma';
+export * from './models';
 export * from './routing';
 export * from './stripe';
